- changed: version to 11 - fixed: issue while bitmap stride was calculated - fixed: width and height were not read correctly for bitmap of type Windows V4 - fixed: some compilation warnings

+++ SDK/CSR_Mdl.c
@@ -505,7 +505,7 @@ int csrMDLReadSkin(const CSR_Buffer*    pBuffer,
     #endif
 
     // create memory for texture
-    pSkin->m_pData = (unsigned char*)malloc((size_t)(pSkin->m_TexLen * pSkin->m_Count));
+    pSkin->m_pData = (unsigned char*)malloc((size_t)pSkin->m_TexLen * (size_t)pSkin->m_Count);
 
     // read texture from buffer. NOTE 8 bit array, same in all endianness
     return csrBufferRead(pBuffer, pOffset, pSkin->m_TexLen, pSkin->m_Count, pSkin->m_pData);
@@ -677,13 +677,13 @@ int csrMDLReadFrameGroup(const CSR_Buffer*        pBuffer,
         // the read bytes are inverted and should be swapped if the target system is big endian
         if (csrMemoryEndianness() == CSR_E_BigEndian)
             // iterate through time values to swap
-            for (i = 0; i < pFrameGroup->m_Count; ++i)
+            for (i = 0; i < (int)pFrameGroup->m_Count; ++i)
                 // swap the value in the memory (thus 0xAABBCCDD will become 0xDDCCBBAA)
                 csrMemorySwap(&pFrameGroup->m_pTime[i], sizeof(float));
     #endif
 
     // read the frames
-    for (i = 0; i < pFrameGroup->m_Count; ++i)
+    for (i = 0; i < (int)pFrameGroup->m_Count; ++i)
         if (!csrMDLReadFrame(pBuffer, pOffset, pHeader, &pFrameGroup->m_pFrame[i]))
             return 0;
 
@@ -824,7 +824,7 @@ void csrMDLPopulateModel(const CSR_MDLHeader*        pHeader,
         return;
 
     // iterate through sub-frames contained in group
-    for (i = 0; i < pFrameGroup->m_Count; ++i)
+    for (i = 0; i < (int)pFrameGroup->m_Count; ++i)
     {
         // create the vertex buffers required for the sub-frames
         pModel->m_pMesh[i].m_Count = 1;
@@ -933,7 +933,7 @@ void csrMDLReleaseObjects(CSR_MDLHeader*       pHeader,
             if (pFrameGroup[i].m_pFrame)
             {
                 // release frame vertices
-                for (j = 0; j < pFrameGroup[i].m_Count; ++j)
+                for (j = 0; j < (int)pFrameGroup[i].m_Count; ++j)
                     free(pFrameGroup[i].m_pFrame[j].m_pVertex);
 
                 // release frame
